+++ src/__mocks__/index.ts
@@ -1,3 +1,4 @@
 export * from './accountConfig';
+export * from './data';
 export * from './server';
 export * from './utils';