+++ src/lib/sdkDappUI.components.ts
@@ -1,0 +1,1 @@
+export { MvxLedgerConnect } from '@multiversx/sdk-dapp-ui/web-components/mvx-ledger-connect';