+++ src/managers/ToastManager/helpers/createToastsFromTransactions.ts
@@ -18,21 +18,15 @@ interface CreateToastsFromTransactionsReturnType {
 }
 
 interface CreateToastsFromTransactionsParamsType {
-  existingCompletedTransactions?: ITransactionToast[];
-  skipFetchingTransactions?: boolean;
   store?: StoreType;
 }
 
-export async function createToastsFromTransactions({
-  existingCompletedTransactions = [],
-  skipFetchingTransactions = false,
-  ...props
-}: CreateToastsFromTransactionsParamsType): Promise<CreateToastsFromTransactionsReturnType> {
-  const store = props.store ?? getState();
+export async function createToastsFromTransactions(
+  props?: CreateToastsFromTransactionsParamsType
+): Promise<CreateToastsFromTransactionsReturnType> {
+  const store = props?.store ?? getState();
   const pendingTransactionToasts: ITransactionToast[] = [];
-  const completedTransactionToasts: ITransactionToast[] = [
-    ...existingCompletedTransactions
-  ];
+  const completedTransactionToasts: ITransactionToast[] = [];
   const toastList = toastsSliceSelector(store);
   const transactionsSessions = transactionsSliceSelector(store);
   const address = addressSelector(store);
@@ -48,7 +42,6 @@ export async function createToastsFromTransactions({
 
     const interprettedTransactions = await mapServerTransactionsToListItems({
       transactions,
-      skipFetchingTransactions,
       store
     });